+++ cardinal-components-world/src/testmod/java/org/ladysnake/cca/test/world/AmbientVita.java
@@ -77,8 +77,8 @@ public abstract class AmbientVita extends BaseVita implements AutoSyncedComponen
         }
 
         @Override
-        public boolean isSyncOptional() {
-            return true;
+        public boolean isRequiredOnClient() {
+            return false;
         }
 
         @Override